+++ hbw/lib/hbw/logger.rb
@@ -1,4 +1,5 @@
 HBW_FORMAT = ' [HBW] %s'
+HBW_DURATION = 'DURATION'
 
 class HBWFormatterProduction < Logger::Formatter
   include ActiveSupport::TaggedLogging::Formatter
@@ -36,6 +37,19 @@ module HBW
       logger = HBWLogger.new(ActiveSupport::Logger.new(STDOUT))
       logger.level = ENV['LOG_LEVEL'] || 'info'
       logger
+    end
+
+    def log_duration(level, start)
+      tagged_logger = ActiveSupport::TaggedLogging.new(logger)
+
+      tagged_logger.tagged(HBW_DURATION) do
+        tagged_logger.public_send(level) do
+          format('Completed %s "%s" in %0.2fms',
+                 request.method,
+                 request.original_fullpath,
+                 (Time.now.to_f - start) * 1000)
+        end
+      end
     end
   end
 end